JavaDoc + better serialization of DelayRange + renaming of FeaturedImageProcessor to FeaturedImageResolver.
