+++ crawler/web/src/main/java/com/norconex/crawler/web/doc/operations/delay/impl/BaseDelayResolverConfig.java
@@ -38,37 +38,22 @@ import lombok.experimental.Accessors;
  * in order from the best behaved to the least.
  * </p>
  * <ul>
- *   <li><b>crawler</b>: the delay is applied between each URL download
+ *   <li><b>CRAWLER</b>: the delay is applied between each URL download
  *       within a crawler instance, regardless how many threads are defined
  *       within that crawler, or whether URLs are from the
  *       same site or not.  This is the default scope.</li>
- *   <li><b>site</b>: the delay is applied between each URL download
+ *   <li><b>SITE</b>: the delay is applied between each URL download
  *       from the same site within a crawler instance, regardless how many
  *       threads are defined. A site is defined by a URL protocol and its
  *       domain (e.g. http://example.com).</li>
- *   <li><b>thread</b>: the delay is applied between each URL download from
+ *   <li><b>THREAD</b>: the delay is applied between each URL download from
  *       any given thread.  The more threads you have the less of an
  *       impact the delay will have.</li>
  * </ul>
- * <h3>
- * XML configuration usage:
- * </h3>
- * <p>
- * The following should be shared across concrete implementations
- * (which can add more configurable attributes and tags).
- * </p>
- * {@nx.xml
- * <delay class="(implementing class)"
- *     default="(milliseconds)"
- *     ignoreRobotsCrawlDelay="[false|true]"
- *     scope="[crawler|site|thread]">
- * </delay>
- * }
  * @since 2.5.0
  */
 @Data
 @Accessors(chain = true)
-@SuppressWarnings("javadoc")
 public class BaseDelayResolverConfig {
 
     public enum DelayResolverScope {
@@ -81,25 +66,19 @@ public class BaseDelayResolverConfig {
             DelayResolverScope.CRAWLER;
 
     /**
-     * The default delay in milliseconds.
-     * @param defaultDelay default deleay
-     * @return default delay
+     * The default delay.
      */
     private Duration defaultDelay = DEFAULT_DELAY;
 
     /**
      * Whether to ignore crawl delays specified in a site robots.txt
-     * file.  Not applicable when robots.txt are ignored.
-     * @param ignoreRobotsCrawlDelay <code>true</code> if ignoring
-     *            robots.txt crawl delay
-     * @return <code>true</code> if ignoring robots.txt crawl delay
+     * file. Not applicable when robots.txt are ignored.
      */
     private boolean ignoreRobotsCrawlDelay = false;
 
     /**
-     * Gets the delay scope.
-     * @param scope one of "crawler", "site", or "thread".
-     * @return delay scope
+     * Gets the delay scope. See class documentation for a description
+     * of supported scopes.
      */
     private DelayResolverScope scope = DEFAULT_SCOPE;
 }
