+++ crawler/web/src/main/java/com/norconex/crawler/web/doc/operations/delay/impl/CrawlerDelay.java
@@ -18,9 +18,14 @@ import org.apache.commons.lang3.mutable.MutableLong;
 
 import com.norconex.commons.lang.Sleeper;
 
+import lombok.EqualsAndHashCode;
+import lombok.ToString;
+
 /**
  * It is assumed there will be one instance of this class per crawler defined.
  */
+@EqualsAndHashCode(onlyExplicitlyIncluded = true)
+@ToString(onlyExplicitlyIncluded = true)
 public class CrawlerDelay extends AbstractDelay {
 
     private MutableLong lastHitEpochNanos = new MutableLong(-1);
