+++ crawler/web/src/main/java/com/norconex/crawler/web/doc/operations/delay/impl/DelayRange.java
@@ -18,11 +18,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
+import lombok.experimental.FieldNameConstants;
 
 @Data
 @Accessors(chain = true)
 @AllArgsConstructor
 @NoArgsConstructor
+@FieldNameConstants
 public class DelayRange<T> {
     private T start;
     private T end;
