+++ crawler/web/src/main/java/com/norconex/crawler/web/doc/operations/delay/impl/GenericDelayResolver.java
@@ -25,7 +25,6 @@ import java.util.List;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.norconex.commons.lang.CircularRange;
 import com.norconex.commons.lang.event.EventListener;
-import com.norconex.commons.lang.time.DurationParser;
 import com.norconex.crawler.core.event.CrawlerEvent;
 
 import lombok.EqualsAndHashCode;
@@ -67,41 +66,6 @@ import lombok.ToString;
  *       any given thread.  The more threads you have the less of an
  *       impact the delay will have.</li>
  * </ul>
- *
- * <p>
- * As of 2.7.0, XML configuration entries expecting millisecond durations
- * can be provided in human-readable format (English only), as per
- * {@link DurationParser} (e.g., "5 minutes and 30 seconds" or "5m30s").
- * </p>
- *
- * {@nx.xml.usage
- * <delay class="com.norconex.crawler.web.delay.impl.GenericDelayResolver"
- *       default="(milliseconds)"
- *       ignoreRobotsCrawlDelay="[false|true]"
- *       scope="[crawler|site|thread]">
- *   <schedule
- *       dayOfWeek="from (week day) to (week day)"
- *       dayOfMonth="from [1-31] to [1-31]"
- *       time="from (HH:mm) to (HH:mm)">
- *     (delay in milliseconds)
- *   </schedule>
- *
- *   (... repeat schedule tag as needed ...)
- * </delay>
- * }
- *
- * {@nx.xml.example
- * <delay class="GenericDelayResolver"
- *     default="5 seconds" ignoreRobotsCrawlDelay="true" scope="site" >
- *   <schedule dayOfWeek="from Saturday to Sunday">1 second</schedule>
- * </delay>
- * }
- *
- * <p>
- * The above example set the minimum delay between each document download
- * on a given site to 5 seconds, no matter what the crawler robots.txt may
- * say, except on weekend, where it is more agressive (1 second).
- * </p>
  */
 @EqualsAndHashCode
 @ToString
