+++ crawler/web/src/main/java/com/norconex/crawler/web/doc/operations/image/impl/FeaturedImageResolver.java
@@ -14,11 +14,11 @@
  */
 package com.norconex.crawler.web.doc.operations.image.impl;
 
-import static com.norconex.crawler.web.doc.operations.image.impl.FeaturedImageProcessorConfig.COLLECTOR_FEATURED_IMAGE_INLINE;
-import static com.norconex.crawler.web.doc.operations.image.impl.FeaturedImageProcessorConfig.COLLECTOR_FEATURED_IMAGE_PATH;
-import static com.norconex.crawler.web.doc.operations.image.impl.FeaturedImageProcessorConfig.COLLECTOR_FEATURED_IMAGE_URL;
-import static com.norconex.crawler.web.doc.operations.image.impl.FeaturedImageProcessorConfig.DEFAULT_IMAGE_CACHE_DIR;
-import static com.norconex.crawler.web.doc.operations.image.impl.FeaturedImageProcessorConfig.DEFAULT_STORAGE_DISK_DIR;
+import static com.norconex.crawler.web.doc.operations.image.impl.FeaturedImageResolverConfig.DEFAULT_IMAGE_CACHE_DIR;
+import static com.norconex.crawler.web.doc.operations.image.impl.FeaturedImageResolverConfig.DEFAULT_STORAGE_DISK_DIR;
+import static com.norconex.crawler.web.doc.operations.image.impl.FeaturedImageResolverConfig.FEATURED_IMAGE_INLINE_FIELD;
+import static com.norconex.crawler.web.doc.operations.image.impl.FeaturedImageResolverConfig.FEATURED_IMAGE_PATH_FIELD;
+import static com.norconex.crawler.web.doc.operations.image.impl.FeaturedImageResolverConfig.FEATURED_IMAGE_URL_FIELD;
 import static java.util.Optional.ofNullable;
 import static org.apache.commons.lang3.StringUtils.endsWithIgnoreCase;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
@@ -60,8 +60,8 @@ import com.norconex.crawler.core.event.listeners.CrawlerLifeCycleListener;
 import com.norconex.crawler.core.fetch.FetchResponse;
 import com.norconex.crawler.core.fetch.Fetcher;
 import com.norconex.crawler.web.doc.WebCrawlDocContext;
-import com.norconex.crawler.web.doc.operations.image.impl.FeaturedImageProcessorConfig.Storage;
-import com.norconex.crawler.web.doc.operations.image.impl.FeaturedImageProcessorConfig.StorageDiskStructure;
+import com.norconex.crawler.web.doc.operations.image.impl.FeaturedImageResolverConfig.Storage;
+import com.norconex.crawler.web.doc.operations.image.impl.FeaturedImageResolverConfig.StorageDiskStructure;
 import com.norconex.crawler.web.fetch.HttpFetchRequest;
 import com.norconex.crawler.web.fetch.HttpFetcher;
 import com.norconex.crawler.web.fetch.HttpMethod;
@@ -75,7 +75,7 @@ import lombok.extern.slf4j.Slf4j;
 /**
  * <p>
  * Document processor that extract the "main" image from HTML pages.
- * Since HTML is expected, this class should only be used at
+ * Since HTML is expected, this class should only be used as a
  * pre-import processor. It is possible for this processor to not find any
  * image.
  * </p>
@@ -93,135 +93,21 @@ import lombok.extern.slf4j.Slf4j;
  *
  * <h3>Storing the image</h3>
  * <p>
- * One or more storage method can be specified. Here are
- * the possible storage options:
+ * When identified, the featured image can be stored either on local disk,
+ * or as a metadata field in Base64 format, or simply as a URL pointing
+ * to its remote location. See {@link FeaturedImageResolverConfig} for details.
  * </p>
- * <ul>
- *   <li>
- *     <b>url</b>: Default. The absolute image URL is stored in a
- *     <code>collector.featured-image-url</code> field.
- *     When only this option is set, scaling options and image format
- *     have no effect.
- *   </li>
- *   <li>
- *     <b>inline</b>: Stores a Base64 string of the scaled image, in the format
- *     specified, in a <code>collector.featured-image-inline</code> field.
- *     The string is ready to be
- *     used inline, in a &lt;img src="..."&gt; tag.
- *   </li>
- *   <li>
- *     <b>disk</b>: Stores the scaled image on the file system, in the format
- *     and directory specified. A reference to the file on disk is stored
- *     in a <code>collector.featured-image-path</code> field.
- *   </li>
- * </ul>
- *
- * {@nx.xml.usage
- * <processor class="com.norconex.crawler.web.doc.operations.image.impl.FeaturedImageProcessor">
- *
- *    <pageContentTypePattern>
- *        (Optional regex to overwrite default matching of HTML pages)
- *    </pageContentTypePattern>
- *
- *    <domSelector>
- *        (Optional CSS-like path matching one or more image elements)
- *    </domSelector>
- *    <minDimensions>
- *        (Minimum pixel size for an image to be considered.
- *         Default is 400x400).
- *    </minDimensions>
- *    <largest>[false|true]</largest>
- *
- *    <imageCacheSize>
- *        (Maximum number of images to cache for faster processing.
- *         Set to 0 to disable caching.)
- *    </imageCacheSize>
- *    <imageCacheDir>
- *        (Directory where to create the image cache)
- *    </imageCacheDir>
- *
- *    <storage>
- *        [url|inline|disk]
- *        (One or more, comma-separated. Default is "url".)
- *    </storage>
- *
- *    <!-- Only applicable for "inline" and "disk" storage: -->
- *    <scaleDimensions>
- *        (Target pixel size the featured image should be scaled to.
- *         Default is 150x150.)
- *    </scaleDimensions>
- *    <scaleStretch>
- *        [false|true]
- *        (Whether to stretch to match scale size. Default keeps aspect ratio.)
- *    </scaleStretch>
- *    <scaleQuality>
- *        [auto|low|medium|high|max]
- *        (Default is "auto", which tries the best balance between quality
- *         and speed based on image size. The lower the quality the faster
- *         it is to scale images.)
- *    </scaleQuality>
- *    <imageFormat>
- *        (Target format of stored image. E.g., "jpg", "png", "gif", "bmp", ...
- *         Default is "png")
- *    </imageFormat>
- *
- *    <!-- Only applicable for "disk" storage: -->
- *    <storageDiskDir structure="[url2path|date|datetime]">
- *        (Path to directory where to store images on disk.)
- *    </storageDiskDir>
- *    <storageDiskField>
- *        (Overwrite default field where to store the image path.
- *         Default is {@value #COLLECTOR_FEATURED_IMAGE_PATH}.)
- *    </storageDiskField>
- *
- *    <!-- Only applicable for "inline" storage: -->
- *    <storageInlineField>
- *        (Overwrite default field where to store the inline image.
- *         Default is {@value #COLLECTOR_FEATURED_IMAGE_INLINE}.)
- *    </storageInlineField>
- *
- *    <!-- Only applicable for "url" storage: -->
- *    <storageUrlField>
- *        (Overwrite default field where to store the image URL.
- *         Default is {@value #COLLECTOR_FEATURED_IMAGE_URL}.)
- *    </storageUrlField>
- *
- * </processor>
- * }
- *
- * When specifying an image size, the format is <code>[width]x[height]</code>
- * or a single value. When a single value is used, that value represents both
- * the width and height (i.e., a square).
- *
- * {@nx.xml.example
- * <preImportProcessors>
- *   <processor class="FeaturedImageProcessor">
- *     <minDimensions>300x400</minDimensions>
- *     <scaleDimensions>50</scaleDimensions>
- *     <imageFormat>jpg</imageFormat>
- *     <scaleQuality>max</scaleQuality>
- *     <storage>inline</storage>
- *   </processor>
- * </preImportProcessors>
- * }
- * <p>
- * The above example extracts the first image being 300x400 or larger, scaling
- * it down to be 50x50 and storing it as an inline JPEG in a document field,
- * preserving aspect ratio and using the best quality possible.
- * </p>
- *
  * @since 2.8.0
  */
-@SuppressWarnings("javadoc")
 @EqualsAndHashCode
 @ToString
 @Slf4j
-public class FeaturedImageProcessor
+public class FeaturedImageResolver
         extends
         CrawlerLifeCycleListener
         implements
         DocumentConsumer,
-        Configurable<FeaturedImageProcessorConfig> {
+        Configurable<FeaturedImageResolverConfig> {
 
     //TODO add ability to extract from popular HTML <meta> for
     // featured image
@@ -229,8 +115,8 @@ public class FeaturedImageProcessor
     //TODO add option to process embedded images (base 64)
 
     @Getter
-    private final FeaturedImageProcessorConfig configuration =
-            new FeaturedImageProcessorConfig();
+    private final FeaturedImageResolverConfig configuration =
+            new FeaturedImageResolverConfig();
 
     private static final Map<Path, ImageCache> IMG_CACHES = new HashMap<>();
 
@@ -258,7 +144,8 @@ public class FeaturedImageProcessor
         // Initialize image cache directory
         if (configuration.getImageCacheSize() > 0) {
             resolvedImageCacheDir = ofNullable(configuration.getImageCacheDir())
-                    .orElseGet(() -> workDir.resolve(DEFAULT_IMAGE_CACHE_DIR));
+                    .orElseGet(() -> workDir.resolve(
+                            DEFAULT_IMAGE_CACHE_DIR));
             try {
                 Files.createDirectories(resolvedImageCacheDir);
                 LOG.info(
@@ -275,7 +162,7 @@ public class FeaturedImageProcessor
         }
 
         // Initialize image directory
-        if (configuration.getStorage().contains(Storage.DISK)) {
+        if (configuration.getStorages().contains(Storage.DISK)) {
             resolvedStorageDiskDir = ofNullable(
                     configuration.getStorageDiskDir()).orElseGet(
                             () -> workDir.resolve(DEFAULT_STORAGE_DISK_DIR));
@@ -338,21 +225,21 @@ public class FeaturedImageProcessor
     private void storeImage(FeaturedImage img, Doc doc)
             throws IOException {
         var imgFormat = configuration.getImageFormat();
-        if (configuration.getStorage().contains(Storage.URL)) {
+        if (configuration.getStorages().contains(Storage.URL)) {
             doc.getMetadata().add(
                     Objects.toString(
                             configuration.getStorageUrlField(),
-                            COLLECTOR_FEATURED_IMAGE_URL),
+                            FEATURED_IMAGE_URL_FIELD),
                     img.getUrl());
         }
-        if (configuration.getStorage().contains(Storage.INLINE)) {
+        if (configuration.getStorages().contains(Storage.INLINE)) {
             doc.getMetadata().add(
                     Objects.toString(
                             configuration.getStorageInlineField(),
-                            COLLECTOR_FEATURED_IMAGE_INLINE),
+                            FEATURED_IMAGE_INLINE_FIELD),
                     img.toHTMLInlineString(imgFormat));
         }
-        if (configuration.getStorage().contains(Storage.DISK)) {
+        if (configuration.getStorages().contains(Storage.DISK)) {
             Path imageFile = null;
             if (configuration
                     .getStorageDiskStructure() == StorageDiskStructure.DATE) {
@@ -390,14 +277,14 @@ public class FeaturedImageProcessor
             doc.getMetadata().add(
                     Objects.toString(
                             configuration.getStorageDiskField(),
-                            COLLECTOR_FEATURED_IMAGE_PATH),
+                            FEATURED_IMAGE_PATH_FIELD),
                     imageFile.toFile().getAbsolutePath());
         }
     }
 
     private boolean savingImage() {
-        return configuration.getStorage().contains(Storage.INLINE)
-                || configuration.getStorage().contains(Storage.DISK);
+        return configuration.getStorages().contains(Storage.INLINE)
+                || configuration.getStorages().contains(Storage.DISK);
     }
 
     private FeaturedImage findFeaturedImage(
